src/couchbase-init/requirements.txt
couchbase==4.3.2
PyYAML==6.0.1
